works with webpack
