+++ .gitignore
@@ -1,0 +1,20 @@
+FFmpeg/
+Makefile
+cache-wasm/
+config.h
+doc/
+ffbuild/
+libavcodec/
+libavdevice/
+libavfilter/
+libavformat/
+libavutil/
+/.version
+/libswresample/
+/libswscale/
+src
+wasm/node_modules/
+wasm/package-lock.json
+wasm/package.json
+wasm/yarn-error.log
+lzo-*
