+++ build-lzo-with-docker.sh
@@ -1,10 +1,10 @@
 #!/bin/bash -x
 
-EM_VERSION=1.39.18-upstream
+EM_VERSION=2.0.18
 
-docker pull trzeci/emscripten:$EM_VERSION
+docker pull emscripten/emsdk:$EM_VERSION
 docker run \
   -v $PWD:/src \
   -v $PWD/cache-wasm:/emsdk_portable/.data/cache/wasm \
-  trzeci/emscripten:$EM_VERSION \
+  emscripten/emsdk:$EM_VERSION \
   sh -c 'bash ./build-lzo.sh'
