+++ wasm/main.html
@@ -3,8 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <title></title>
-  <script src="./dist/lzo-wasm.js"></script>
-  <script src="./lzo.js"></script>
+  <script src="./test.js" type="module"></script>
 </head>
 <body>
   <input type="file" id="file-selector">
